[SYCL] move sycl device mgmt to separate header
